ResultData에 data1Name 추가

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -44,7 +44,7 @@ public class UsrArticleController {
     ResultData<Integer> writeArticleRd = articleService.writeArticle(loginedMemberId,title,body);
     int id = writeArticleRd.getData1();
     Article article = articleService.getArticle(id);
-    return ResultData.from(writeArticleRd.getResultCode(),writeArticleRd.getMsg(),article);
+    return ResultData.newData(writeArticleRd,"article",article);
   }
 
 
@@ -55,7 +55,7 @@ public class UsrArticleController {
     if( article == null){
       return ResultData.from("F-1", Ut.f("%d번 게시물이 존재하지 않습니다.",id));
     }
-    return ResultData.from("S-1", Ut.f("%d번 게시물입니다.",id));
+    return ResultData.from("S-1", Ut.f("%d번 게시물입니다.",id),"article",article);
   }
 
 
@@ -63,31 +63,43 @@ public class UsrArticleController {
   @ResponseBody
   public ResultData<List<Article>> getArticles() {
     List<Article> articles = articleService.getArticles();
-    return ResultData.from("S-1","게시물 리스트입니다.",articles);
+    return ResultData.from("S-1","게시물 리스트입니다.","articles",articles);
   }
   @RequestMapping("/usr/article/doDelete")
   @ResponseBody
-  public ResultData<Integer> doDelete(int id) {
+  public ResultData<Integer> doDelete(HttpSession httpSession,int id) {
     Article article = articleService.getArticle(id);
+    int loginedMemberId = 0;
+    boolean isLogined = false;
 
+    if(httpSession.getAttribute("loginMemberId")!=null){
+      isLogined = true;
+      loginedMemberId = (int) httpSession.getAttribute("logindMemberId");
+    }
+    if(isLogined == false){
+      return ResultData.from("F-A","로그인 후 이용하세요");
+    }
+    if(article.getMemberId() != loginedMemberId){
+      return ResultData.from("F-2","권한이 없습니다.");
+    }
     if( article == null){
       return  ResultData.from("F-1", Ut.f("%d번 게시물이 존재하지 않습니다.",id));
     }
     articleService.deleteArticle(id);
 
-    return ResultData.from("S-1", Ut.f("%d번 게시물을 삭제했다.",id),id);
+    return ResultData.from("S-1", Ut.f("%d번 게시물을 삭제했다.",id),"id",id);
   }
 
   @RequestMapping("/usr/article/doModify")
   @ResponseBody
-  public ResultData<Integer> doModify(int id,String title,String body) {
+  public ResultData<Integer> doModify(HttpSession httpSession, int id,String title,String body) {
     Article article = articleService.getArticle(id);
 
     if( article == null){
       return ResultData.from("F-1", Ut.f("%d번 게시물이 존재하지 않습니다.",id));
     }
     articleService.modifyArticle(id,title,body);
-    return  ResultData.from("S-1", Ut.f("%d번 게시물을 수정했다.",id),id);
+    return  ResultData.from("S-1", Ut.f("%d번 게시물을 수정했다.",id),"id",id);
   }
